status memory changed greatly
in last commit, yesterday rewrite the status memory, but there were
timing issues remained. Today fixed timing issue and test it. In recall
mode, the result is the same as the original version. compared waveform
of the two version, new version neuron generate spikes the same as
original.
Learn mode need to be tested.
StatusMem_new is removed, code in this file is copied to StatusMem and
then fixed.
StatusMem_original is the old version, Copied from StatusMem. It works
fine but cannot be synthesized.